Use control character in sed lines with variables
A cookie value with a '/' in it caused a sed error during postinst;

```
sed: -e expression #1, char 53: unknown option to `s'
```

We use a control character (RS - record separator) instead of / to
reduce the chances of a collision with a valid cookie string.

I applied the same change to nodename even though a / in node name
would be an error, for consistency.