Exclude PPC64LE arch for packaging
There is an `Unknown privilege violation (03)` error emitted to the
log followed by a register dump right before gradle does it's
download. Then everything seems to continue as normal, but suddenly
the linter fails. This happens on PPC64LE and it seems to work for
other arches. So let's exclude the combination which fail for now.

See issue: https://github.com/apache/couchdb-pkg/issues/140