RPM support for nouveau
This is a simple version that goes along with the release artifacts
file setup mostly, just adds a systemd unit file for nouveau.

Nouveau config setting can be automatically enabled the rpm install
using `COUCHDB_NOUVEAU_ENABLE=true`.

For instance:

 `COUCHDB_NOUVEAU_ENABLE=true COUCHDB_COOKIE=c00kie dnf install couchdb*.rpm`

Java dependency is declared as a `Suggests:...`. This is the weakest
form of dependency declaration. It won't automatically bring in java
21.

In the next minor version it may become a `Recommends:` which is one
level higher and will automatically install it.

If nouveau config is enabled during the RPM install, user is then
urged to install java (suggestion via dnf but they may install their
own from any vendor) and to enable the systemd unit to auto-start
after boot.

CouchDB service also works this way: it has to be enabled
explicitly. This part didn't change:

 `systemctl enable couchdb`
 `systemctl enable couchdb-nouveau`

This lets a user choose to start just CouchDB on a server, just
Nouveau, or both CouchDB and Noveau.

In addition, since we're updating the RPM package, remove some of
deprecated/redundant settings needed for old CentOS-es:

  - Log to journald instead of a file
  - Log level is already set to info, so no need to override it again in the RPM package
  - No need for an init file, all supported OSes have systemd