Remove buggy 10-nouveau.ini creation
In an effort to make nouveau easier to use the couchdb-nouveau package would
attempt to reconfigure couchdb. However this fails if you install couchdb and
couchdb-nouveau at the same time (and doing so is quite normal).

Remove this code and print a note instead. The sysadm can make the single
`enable = true` configuration change themselves.

closes https://github.com/apache/couchdb-pkg/issues/154